pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "qa-task"
version = "0.1.0"
description = ""
authors = ["Likai Ren <lebsfi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.31.0"
robotframework = "^6.1.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.9.1"
pre-commit = "^3.4.0"
robotframework-lint = "^1.1"
